requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.4
python-dotenv==1.0.0
gspread==5.12.4
oauth2client==4.1.3
